refactor: split display and item
